+++ mex/extractors/voxco/main.py
@@ -4,7 +4,7 @@ from typing import Any
 from dagster import asset
 
 from mex.common.cli import entrypoint
-from mex.common.ldap.transform import transform_ldap_persons_to_mex_persons
+from mex.common.ldap.transform import transform_ldap_persons_to_extracted_persons
 from mex.common.models import (
     ExtractedActivity,
     ExtractedOrganization,
@@ -85,7 +85,7 @@ def extracted_mex_persons_voxco(
     ldap_persons = extract_ldap_persons_voxco(
         [ResourceMapping.model_validate(r) for r in voxco_resource_mappings]
     )
-    mex_persons = transform_ldap_persons_to_mex_persons(
+    mex_persons = transform_ldap_persons_to_extracted_persons(
         ldap_persons,
         extracted_primary_source_ldap,
         extracted_organizational_units,