+++ CHANGELOG.md
@@ -9,13 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- extract mandatory fields for igs variables and variable groups
+- update mex-common dependency to 1.4
+
 ### Changes
+
+- read datenkompass default values from mapping
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- odk variable identifier in primary source
 
 ### Security